+++ cryptography/kzg_multi_open/src/naive.rs
@@ -1,6 +1,8 @@
 use crate::{commit_key::CommitKey, opening_key::OpeningKey};
 use bls12_381::{ff::Field, multi_pairings, G1Point, G1Projective, G2Point, G2Prepared, Scalar};
-use polynomial::monomial::{lagrange_interpolate, poly_eval, poly_sub, vanishing_poly, PolyCoeff};
+use polynomial::poly_coeff::{
+    lagrange_interpolate, poly_eval, poly_sub, vanishing_poly, PolyCoeff,
+};
 
 /// This modules contains code to create and verify opening proofs in a naive way.
 /// It is also general, meaning the points we are creating opening proofs
