+++ cryptography/polynomial/benches/benchmark.rs
@@ -1,6 +1,6 @@
 use bls12_381::Scalar;
 use bls12_381::{ff::Field, group::Group, G1Projective};
-use crate_crypto_internal_eth_kzg_polynomial::{domain::Domain, monomial::poly_eval};
+use crate_crypto_internal_eth_kzg_polynomial::{domain::Domain, poly_coeff::poly_eval};
 use criterion::{criterion_group, criterion_main, Criterion};
 
 pub fn bench_polynomial_evaluation(c: &mut Criterion) {
